server/pyproject.toml
[project]
name = "langgraph-agents"
version = "0.0.1"
description = "Agents for LangGraph"
authors = [
    { name = "Tadas Valaitis", email = "valaitis.t@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11.0,<=3.12.9"
dependencies = [
    "langgraph>=0.2.19",
    "langchain-openai>=0.1.22",
    "langchain-anthropic>=0.1.23",
    "langchain>=0.2.14",
    "langchain-fireworks>=0.1.7",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.2.13",
    "pandas>=2.1.1",
    "numpy>=1.26.1",
    "tabulate>=0.9.0",
    "fastparquet>=2023.10.1",
    "pyarrow>=14.0.1",
    "langchain_chroma>=0.2.3",
    "openevals>=0.0.18",
    "beautifulsoup4>=4.13.3",
    "openevals[e2b-code-interpreter]==0.0.18",
    "backtesting==0.6.1",
    "ta==0.11.0",
    "aiofiles>=23.2.1",
    "pyright>=1.1.400",
    "pandas-ta>=0.3.14b",
    "docker (>=7.1.0,<8.0.0)",
    "ccxt (>=4.4.80,<5.0.0)",
    "krakenex (>=2.2.2,<3.0.0)",
    "langgraph-checkpoint-postgres (>=2.0.21,<3.0.0)",
    "scikit-optimize (>=0.10.2,<0.11.0)",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest-asyncio","langgraph-cli[inmem]","langgraph-api"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["doc_agent", "react_agent", "strategy_agent"]
[tool.setuptools.package-dir]
"doc_agent" = "src/doc_agent"
"react_agent" = "src/react_agent"
"strategy_agent" = "src/strategy_agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
include = ["*.py", "*.pyi", "*.ipynb"]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"
